# Conflicts:
#	Functions.py

+++ chap5methods.py
@@ -39,9 +39,9 @@ def ModifiedEuler(Expression, h, n, bound1, bound2, y0, fix):
     n = bound1+h
     i = int(0)
     while (n <= bound2):
-        k1 = round((h * func2(Expression, matrix[i][0], matrix[i][1])), fix)
-        k2 = round(
-            (h * func2(Expression, matrix[i+1][0], (k1 + (matrix[i][1])))), fix)
+        k1 = float(round((h * func2(Expression, matrix[i][0], matrix[i][1])), fix))
+        k2 = float(round(
+            (h * func2(Expression, matrix[i+1][0], (k1 + (matrix[i][1])))), fix))
 
         matrix[i+1][1] = round((matrix[i][1] + 0.5*(k1 + k2)), fix)
 
@@ -73,11 +73,11 @@ def Huen(Expression, h, n, bound1, bound2, y0, fix):
     n = bound1+h
     i = int(0)
     while (n <= bound2):
-        k1 = round((func2(Expression, matrix[i][0], matrix[i][1])), fix)
-        k2 = round(
-            (func2(Expression, ((h/3)+matrix[i][0]), (((k1*h)/3) + (matrix[i][1])))), fix)
-        k3 = round(func2(
-            Expression, (((2*h)/3)+matrix[i][0]), (((2*k2*h)/3) + (matrix[i][1]))), fix)
+        k1 = float(round((func2(Expression, matrix[i][0], matrix[i][1])), fix))
+        k2 = float(round(
+            (func2(Expression, ((h/3)+matrix[i][0]), (((k1*h)/3) + (matrix[i][1])))), fix))
+        k3 = float(round(func2(
+            Expression, (((2*h)/3)+matrix[i][0]), (((2*k2*h)/3) + (matrix[i][1]))), fix))
 
         matrix[i+1][1] = round((matrix[i][1] + (1/4)*(k1 + (3*k3))*h), fix)
 
@@ -108,7 +108,7 @@ def Euler(Expression, h, n, bound1, bound2, y0, fix):
     n = bound1+h
     i = int(0)
     while (n <= bound2):
-        k1 = func2(Expression, matrix[i][0], matrix[i][1])
+        k1 = float(func2(Expression, matrix[i][0], matrix[i][1]))
         
         matrix[i+1][1] = round((matrix[i][1] + (h*k1)), fix)
 
@@ -140,10 +140,10 @@ def Rk4(Expression, h, n, bound1, bound2, y0, fix):
     n = bound1+h
     i = int(0)
     while (n <= bound2):
-        k1 = round((h * func2(Expression, matrix[i][0], matrix[i][1])), fix)
-        k2 = round((h * func2(Expression, ((h/2)+matrix[i][0]), ((k1/2) + (matrix[i][1])))), fix)
-        k3 = round((h * func2(Expression, ((h/2)+matrix[i][0]), ((k2/2) + (matrix[i][1])))), fix)
-        k4 = round((h * func2(Expression, matrix[i+1][0], (k3 + (matrix[i][1])))), fix)
+        k1 = float(round((h * func2(Expression, matrix[i][0], matrix[i][1])), fix))
+        k2 = float(round((h * func2(Expression, ((h/2)+matrix[i][0]), ((k1/2) + (matrix[i][1])))), fix))
+        k3 = float(round((h * func2(Expression, ((h/2)+matrix[i][0]), ((k2/2) + (matrix[i][1])))), fix))
+        k4 = float(round((h * func2(Expression, matrix[i+1][0], (k3 + (matrix[i][1])))), fix))
 
 
         matrix[i+1][1] = round((matrix[i][1] + (1/6)*(k1 + 2*(k2+k3) + k4)), fix)
